+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 Change Log
 ==========
+1.3.0 (2014-07-08)
+------------------
+1. Added remote view template support @*.html and all.json (compiled);
+2. Adding *.html -> all.json into build script...
+3. Updating doc (how-to) about view templating changes...
+
+
 1.2.2 (2014-07-08)
 -------------------
 1. fixed bug on region scanning in application's main template (#main);